Refactor `useDimensions` to separate file
